requirements.txt
vk-api==11.9.9
requests==2.31.0
google-auth-transport==0.1.0
google-api-python-client==2.86.0
google-auth-oauthlib==0.8.0